feat(scripts): lint CSS with stylelint (#344)
Currently doing nothing because we don't have a set of rules defined
yet, but you can see a couple of sample rules commented out that I used
for testing.

As noted in previous commits, the API is, unfortunately, async only,
which means we need to change a bunch of other places to make them
async-capable.

Some other limitations are noted inline, the most obvious of which is
that stylelint has an autofix functionality that is very loudly labeled
as "experimental":

> Note: It's an experimental feature. It currently does not respect
> special comments for disabling stylelint within sources (e. g. /*
> stylelint-disable */). Autofixing will be applied regardless of these
> comments.
>
> If you're using both these special comments and autofixing, please
> run stylelint twice as a temporary solution. On the first run, some
> violations could be missed, or some violations might be reported
> incorrectly.

Source: https://stylelint.io/user-guide/cli#autofixing-errors

Another wrinkle is that stylelint doesn't report whether a particular
error is autofixable, so our "fixable" counts are all going to be wrong.

Test plan: Add some errors in JS and JSP files in liferay-portal to make
sure those lints still work, then turn on the commented-out stylelint
rules. See output like this:

https://gist.github.com/wincent/3d9f2da9a4eed0eedf8ef2ac5d0bc29e

An excerpt:

    modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/js/FragmentsEditor.es.js
      44:1  error  Unexpected 'debugger' statement.  no-debugger

    modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/view_soy.jsp
      56:1  error  Unexpected 'debugger' statement.  no-debugger
      70:1  error  Unexpected 'debugger' statement.  no-debugger

    modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/css/ColumnOverlayGrid.scss
      12:10  error  Unexpected unit "vh" (unit-whitelist)  unit-whitelist
      16:7  error  Unexpected unit "vh" (unit-whitelist)  unit-whitelist

    ...

    ✖ 300 problems (300 errors, 0 warnings)

    2 of 3 jobs failed