feat: create a facet-theme generator (#419)
This is based on a pattern used in Liferay JS Toolkit.

The idea is to create (low level) "facet" generators that can be
reused, using composition, in other top level generators (in this case
we'll use it for the theme, classic and admin generators).

This generator prompts the user with questions needed to create a
bare-bones theme and writes its associated files.