feat: base the default generator in facet-theme (#419)
We let the facet-theme generator handle all intervention related to
project creation and then delegate on a specialized generator that
creates the theme's custom.css file and tunes the project.

Note that we also remove the tests because the functionality they
were testing has been removed, as we are not supporting CLI args any
more.