feat: base the layout generator in facet-theme (#419)
Because we have removed inheritance from the theme generator, we don't
want the layout generator to extend from it any more so we refactor it.

It is also split into two specialized subgenerators: one to generate a
brand new layout project, and another one to generate a layout inside an
existing theme project.

The layout.ftl file is not needed because the LayoutCreator returns its
content. In previous versions there was a CLI argument to skip layout
creation but it has been removed so there's no point in creating a
placeholder layout file any more.