refactor(scripts): prepare preflight to run multiple checks
This commit was original part of:

https://github.com/liferay/liferay-npm-tools/pull/348

but we didn't proceed with it because we didn't have consensus around
what the file naming conventions should be.

Nevertheless, we can build on this to implement other preflight checks.
Specifically, I want to be able to run some checks against package.json
files. Look at the next commit for that.

Here follows the original commit from #348:

> This is some initial refactoring that will give us a place to enforce
> file-naming conventions. It divides the preflight check into two
> phases:
>
> - Config file checking (ie. the existing check).
> - Source file checking (currently a placeholder, to be implemented
>   once we have locked down the conventions).
>
> https://github.com/liferay/liferay-frontend-guidelines/issues/78