feat: create initial QA scripts
This is based on a pattern used in Liferay JS Toolkit. The idea is to
create a yarn workspace where QA projects are generated that can be
deployed to a Portal instance using the local development packages.

The setup is based on yarn and lerna and assumes that node resolution
will find local packages in the project's workspace so there's no need
to link anything to make the magic happen.