feat(scripts): lint against multiline comments in SCSS files
This requires us to build a custom stylelint plugin/rule because there
doesn't seem to be a built-in rule for this. Note that I am bundling the
plugin inside liferay-npm-scripts itself to avoid the overheading of
having to publish and separately maintained plugin package.

Also note the custom Jest matcher, `toLintStyles`: I'll extract that
into a separate file when/if we add another custom plugin/rule.

Now that we have a "lint/stylelint" subdirectory, in the next commit I'll do a
small refactor to move the stylelint files down into that.

Closes: https://github.com/liferay/liferay-npm-tools/issues/232