chore(scripts): configure some Sass at-rules
These rules are provided by Sass and used in liferay-portal, so should
not actually be considered as "unknown". There are still some others
that Sass provides (eg. `@debug`, `@while` and so on) that are not in
this list, but as there is no usage of them at this point, I am not
adding them:

https://sass-lang.com/documentation/at-rules/mixin