refactor(scripts): move toLintStyles into separate file (#352)
I'm about to add a second custom stylelint rule so I want the matcher to
be available throughout the project. This commit moves it up into a file
of its own and configures Jest to automatically require it.

Note that I did this configuration at both the top-level and inside
liferay-npm-scripts so that `yarn test` will work regardless of where it
is invoked.

Related: https://github.com/liferay/liferay-npm-tools/issues/352