Fixes #459: Point CSS link to correct place
We need to append the @version but also namespace the package name in
case it's necessary.

By default we assume namespace is necessary because it's the usual
config when using bundler v2 (and bundler v1 doesn't support loaders).
We could be smarter and check for the presence of the namespace bundler
plugin instead of using a setting, but I think the benefit is not worth
the effort given how unlikely it is to disable namespacing.