feat!: add import-extensions rule
When I originally created the issue, I was thinking about ".soy.js"
imports, but in the end we may as well apply it to all ".js" imports.
See the documentation in this commit for details on why this is a
reasonable approach for our codebases, and why I think we won't need any
configurability at this time.

Closes: https://github.com/liferay/eslint-config-liferay/issues/137