feat(scripts): expose "prettier" subcommand
As explained in the README, the intent here is to facilitate editor
integration with our customized spin on Prettier. The idea is that
instead of using the `prettier` executable, you can use
`liferay-npm-scripts prettier` as a more-or-less drop-in replacement.

Next steps will involve showing how to actually use this in some
editors, like Vim and VSCode.