fix: make sort-import-destructures work with long lists
The old implementation happened to work with short lists but it made too
many assumptions about the order in which pairs in the list would be
compared. As soon as we tried running this on files with many names
being destructured (like the one in the included regression test), then
the assumption (that we'd only ever call element in list-order) was
invalidated.

V8 actually uses Timsort under the covers
(https://v8.dev/blog/array-sort), but that is besides the point. We
need to make sure the comparator function passed to `sort` behaves
identically regardless of comparison order.