docs(scripts): make it clearer that you are supposed to copy wrapper
Yesterday I had somebody asking for help getting the VSCode wrapper
working, and it turns out they were referencing this path instead of
using the VSCode wrapper:

    path/to/liferay-portal/modules/node_modules/liferay-npm-scripts/src/scripts/prettier.js

Note that the wrapper is committed to the repo but not included in the
NPM package, by design; you're supposed to download it and stick it
somewhere, and it should not depend on having a copy of the
liferay-portal node_modules checked out (they are too easily
cleaned/deleted etc).

So, update the instructions to make it clearer that you are supposed to
download the wrapper and stick it somewhere on your local disk.