refactor(scripts): pick a better name for the debug channel
Nobody would have actually seen this, because there are no `debug` calls
in the wrapper, but when I was trouble-shooting with a colleague
yesterday I had to explain why they might see a "greg" channel in their
console drawer in VSCode. Clearly ridiculous, so we're better off just
changing it.