feat(scripts): catch banned dependencies in package.json files
This is an evolution of this commit:

https://github.com/liferay/liferay-npm-tools/pull/348/commits/60663da20fe7fb7e7443e8b37f393169984e50b4

In that PR, which was never merged, the commit enforced file-naming
conventions. In this one, we're going to do a check against package.json
files instead, to make sure they don't include any banned dependencies.

I asked about doing this in the Java source formatter here:

https://github.com/brianchandotcom/liferay-portal/pull/84226#issuecomment-582815408

but got no reply. It seems preferable for us to do it anyway, so that we
can have control over what checks run and how:

Test plan: Run in liferay-portal; each `getPaths` call takes about 4
seconds on my machine even though that means scanning the entire repo --
we could memoize that or make it "smarter" in CI (ie. by using the
changed paths detection to short-circuit traversal instead of doing it
at the end as a filtering step). Example of running in a project with
bad dependencies:

    package.json: BAD - contains blacklisted dependency: liferay-npm-bundler-loader-css-loader
    package.json: BAD - contains blacklisted dependency: liferay-npm-bundler-loader-style-loader

Example of running from the top level and hitting a file with some other
error (I went in and malformed the JSON); line-break added for
readability only:

    apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/package.json:
    BAD - error thrown during checks: SyntaxError: Unexpected token } in JSON at position 1162