feat(scripts): format JSON with Prettier
Seeing as we are adding semantic checks for package.json files, we may
as well pass the files through Prettier too. That means that we can use
the npmscripts.config.js "check" and "fix" globs to control which files
we want to run against, and it can all just flow through the same
pipeline.

Note that we are being very conservative with our targeting here. The
preset only looks at ".json" files in each project root, and the
suggested configuration change in liferay-portal's top-level
npmscripts.config.js will do the same (note second item below):

    const CHECK_AND_FIX_GLOBS = [
            '/{,dxp/}*.js',
            '/{,dxp/}apps/*/*/*.{js,json}',
            '/{,dxp/}apps/*/*/{src,test}/**/*.{js,scss}',
            '/{,dxp/}apps/*/*/{src}/**/*.{jsp,jspf}'
    ];

I confirmed in this PR that the way Prettier formats the files is not in
conflict with the Java source formatter (only the end-of-file line
endings differ, and Java SF doesn't care about that in
package.json files):

https://github.com/wincent/liferay-portal/pull/131