feat(scripts): enforce trailing commas in JS
This wasn't previously safe to do because IE will choke on a trailing
comma, but now that basically all JS is passing through Babel and
getting transformed away, we can go ahead.

The exceptions to this are theme JS, of which there is almost none, and
JS in JSP, which is not affected (because we adjust Prettier to make
sure that it never adds commas in that context).

Test plan: In a test run over liferay-portal, produces this 77k-line diff:

https://gist.github.com/wincent/2643a647345a98ffb4df9b559aba3688

Contrast that with this, where we skip the special casing of JSP, and it
would produce this 88k-line diff, which would evidently cause problems
in IE:

https://gist.github.com/wincent/223f31e62f74cf542bd8f1b9df087826

Closes: https://github.com/liferay/liferay-npm-tools/issues/380

Related: https://github.com/liferay/liferay-frontend-guidelines/issues/120