feat(scripts): expand default mocks provided in test environment
We've provided some mocks for a while now since #77 was created. At this
point I think we might as well grab some more low-hanging fruit and then
close this one. We can create new issues for specific additions that
we'd like to make in the future.

In this commit I add the bits necessary to do the following cleanup in
liferay-portal:

https://gist.github.com/wincent/e70a525c667d41c75f81a4084a7f2b31

Closes: https://github.com/liferay/liferay-npm-tools/issues/77