chore: update dependencies
Went through the list shown by `yarn outdated` and picked all the ones
that looked "safe". We know from past experience that updating
`@storybook` can be painful, so I left that one as is. Many of the
updates that are "breaking" look innocuous enough, with the exception of
the cosmiconfig update, which requires a code change.