fix(scripts): work around performance regression in Jest 25.1.0
Works around:

https://github.com/facebook/jest/issues/9457

By hard-wiring us to use jsdom v13, which is the last known good version
at this time.

Jest v25.1.0 includes a big jump in the jsdom dependency (from 11.12.0
to 15.2.1), which brings a performance regression bad enough to make the
portal-search-tuning-rankings-web tests time out. More details at:

https://github.com/liferay/liferay-npm-tools/pull/388#issuecomment-587493357

This is a temporary workaround. We should be able to move forward to a
newer version of jsdom once all of the issues with it are fixed.