feat(script)!: sort imports in SCSS files
We sort alphabetically, with blank lines used to divide imports into
groups. We only sort within groups, never across groups, which means
that you can force things to be imported in a certain order if
necessary.

Example output on liferay-portal:

https://gist.github.com/wincent/d5e48a398930b327c9cc29291ee35dc3

Example autofix:

https://gist.github.com/wincent/c9dd73cda61e2c3519a280301ba2458a

Note that in there I inserted a blank-line after an `@import
"variables"` in page_editor just to show how this can be used to prevent
an autosort from being applied; if you look in another file in
page_editor in that diff, you'll see what happens if you don't do that.

Closes: https://github.com/liferay/liferay-npm-tools/issues/389