feat(scripts): disallow explicit .scss extension in imports
This is just a simple implementation of the rule that doesn't try to
handle compound import statements eg.

    @import 'a', 'b', 'c';

because we don't have any of those in liferay-portal. We can add a
separate rule to enforce that that invariant remains true, if we want
to.

Test plan: Copy this over to liferay-portal and see it make these
changes:

https://gist.github.com/wincent/189c5772ae8487b2220a1b638a7bb137

Closes: https://github.com/liferay/liferay-npm-tools/issues/391