chore: update for compliance with current Outbound Licensing Policy
This change is analogous to the one I just made in:

https://github.com/liferay/eslint-config-liferay/pull/151

It updates the license header format to match the latest policy:

https://grow.liferay.com/excellence/Liferay+Outbound+Licensing+Policy

Did this by updating `copyright.js`, running `yarn lint:fix`, and then
applying manual fixes to the three `bin` files (because of their use of
"shebang" lines, the linter adds a redundant copy of the header). I used
a Vim macro to reset the year to its original value (2019) instead of
letting the linter update it to 2020.