chore: update for compliance with current Outbound Licensing Policy
This change is analogous to the one I just made in:

- https://github.com/liferay/eslint-config-liferay/pull/151
- https://github.com/liferay/liferay-npm-tools/pull/394

It updates the license header format to match the latest policy:

https://grow.liferay.com/excellence/Liferay+Outbound+Licensing+Policy

Did this by updating `copyright.js`, preserving the original year,
running `yarn lint:fix`, and then applying manual fixes in the `bin`
files (because of its use of "shebang" line, the linter adds a redundant
copy of the header). I used a Vim macro to strip out the original
headers. Finally, I changed the hardcoded year in the copyright.js
template (2017) to a `<%= YEAR %>`, so that subsequently created files
will use the current year.

Adds a copy of the license to `LICENSES/MIT.txt`.