chore: update for compliance with current Outbound Licensing Policy
This change is analogous to the one I made in:

- https://github.com/liferay/eslint-config-liferay/pull/151
- https://github.com/liferay/liferay-npm-tools/pull/394
- https://github.com/liferay/liferay-js-themes-toolkit/pull/441
- https://github.com/liferay/clay/pull/2947
- https://github.com/liferay/alloy-editor/pull/1382
- https://github.com/liferay/liferay-amd-loader/pull/227

It updates the license header format to match the latest policy:

https://grow.liferay.com/excellence/Liferay+Outbound+Licensing+Policy

Did this by updating `copyright.js` and running `yarn lint:fix`. I used
a Vim macro to strip out the original headers and preserve the original
copyright dates.

Adds a copy of the license to `LICENSES/LGPL-3.0-or-later.txt` as
required by the policy too.