fix: don't break when using @typescript-eslint/parser
The sort-import-destructures rule was working fine in liferay-portal and
in our local test suite, but I discovered while updating the version of
eslint-config-liferay used in Clay to v19.0.1 that the rule's autofixer
was mangling the source because the `start` and `end` properties are
undefined when using the `@typescript-eslint/parser` parser.

`range[0]` and `range[1]`, however, are equivalent, and are set by both
`@typescript-eslint/parser` and the standard parser.