feat: use trailing commas in JSP too
Turns out my prior analysis was based on a misremembering.

This MDN page:

    https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Trailing_commas

claims that we can use trailing commas in IE9 and above, so I was wrong
about this not being safe to put in our JSPs. This page:

    https://2ality.com/2013/07/trailing-commas.html

confirms.

Evidently, I was thinking of this problem:

    https://github.com/brianchandotcom/liferay-portal/pull/79731#issuecomment-543540235

in JSP, which is precisely the opposite of what I was worried about; ie.
that the *lack* of trailing commas in our JSP was actually *causing* us
problems when it came to conditional execution vs formatting.