feat: #516 Use .js as config file
Instead of using .npmbundlerrc, we start using
liferay-npm-bundler.config.js.

This serves two purposes:

1) Let developers customize webpack config
2) Force users to migrate their .npmbundlerrc when moving from bundler
   2.x to 3.x so that they have to inspect its contents and planify how
   to make the migration.

This commit also removes the presets feature. This is because once we
load the config from a .js file, it's trivial to load or combine as many
presets as you want by simply requiring them in the config file and
manipulatings the objects they return. Thus, we leave the "presets" to
the user's side.

Note that the general contract of the liferay-npm-bundler.config.js file
is that it must be a valid CommonJS module which only exports an object
containing the configuration.