fix(changelog): avoid duplicate entries in logs
Because `git-log` with the `-m` switch may report the same commit twice
(if both sides of the merge ended up making changes to the files). An
example of this can be seen in this gist:

https://gist.github.com/wincent/05d360c294c657aea54b474839d561e4

The fix is to use a `Map`, keyed on the commit hash to ensure
uniqueness.