feat(changelog): copy breaking changes into a separate section
The idea is to maximize the visibility of breaking changes by collating
them together at the top. Note that the messages are repeated there
instead of moved there, because otherwise we lose the metadata about
what they actually are (features, bug fixes etc).

The other thing I tried was leaving them inline and marking them with a
":warning:" emoji, but that didn't look very clean.

Test plan: Regen the liferay-npm-scripts changelog with these results:

https://gist.github.com/wincent/4fef71c6899f0a816abc58e8d8b688ec