feat: #520 Implement imports
We inject ImportsLoader into webpack's build to post process every bundle and
substitute require() calls pointint to imported modules by __REQUIRE__() so that
webpack doesn't resolve them and we can, after webpack is run, restore them to
require() calls that will be resolved by Liferay AMD Loader.

Also, refactor the adaptation phase to break it in smaller modules so that it is
manageable.

We are using AST parsers to modify the require() calls. I was tempted to use
regular expressions to make it more efficient but unless it is really needed
we'll go with the more robust option first.