chore: Delegate build and watch to tsc
This is to make spurious build problems disappear.

We don't need specialized scripts for these tasks any more, as tsc
project references feature does the trick.

Build is a bit slower than with our specialized script, but not so
slower as to justify staying out of the blessed way of building
Typescript projects.