fix(scripts): make check/fix globs apply consistently at project level
@julien reported that this PR:

    https://github.com/brianchandotcom/liferay-portal/pull/85713

was failing SF checks in CI but not locally.

The cause was that we had conservative default globs defined at the
project level:

    https://github.com/liferay/liferay-npm-tools/blob/29545e04a57b35b17903aa638c7db2be129995ae/packages/liferay-npm-scripts/src/presets/standard/index.js#L10

so if you run `yarn format` locally from a project, as Julien was, it
won't actually look at any JS files outside of "src/" or "test/". This
has been the case since the beginnings of the preset:

    https://github.com/liferay/liferay-npm-tools/blob/dc58feb8260e3d71b1546782672d5acc08fd6197/packages/liferay-npm-scripts-preset-standard/index.js#L19

(In fact, back then, it was even more conservative, because it only
looked at "src/", not "test/").

Compare that with the top-level globs in liferay-portal:

    https://github.com/liferay/liferay-portal/blob/a70a80655148ff04e8f939bbdc1fa7b46e74e102/modules/npmscripts.config.js#L17

which will catch dotfiles like `.eslintrc.js`. So CI was failing because
wit runs project-wide checks.

I think we can make the globs more aggressive in the preset at this
point to avoid this kind of confusion. The current behavior is
inconsistent enough that I think we can justify calling it a bug.