docs: update release instructions
I just noticed while preparing the release of liferay-npm-scripts
v28.0.2 that I hadn't included the full path the generator; ie.

    ../liferay-changelog-generator/bin/liferay-changelog-generator.js

Let's just simplify the instructions and tell people to use `npx`
because that is much easier. We can assume that if you are working on
the changelog generator itself and you want to use an unreleased version
of the generator then you probably know when not to use `npx` and
provide a full path instead.