chore: #518 Let snapshot releases point to snapshots
The release:snapshot script checks for a clean working copy before
release but in the case where some projects depend on others, we may
need to release snapshots with locally modified package.json files.

For example, if we are releasing a snapshot of liferay-npm-bundler that
must point to a snapshot of liferay-npm-build-tools-common, the way to
do it is modifying the package.json of liferay-npm-bundler prior to
releasing.