feat: use version-tag-prefix automatically
The idea of this change is to make it easier to generate changelogs in a
monorepo with per-package changelogs. At the moment, you have to provide
the `--version` flag in the format that corresponds to the version that
you are going to release (eg. `--version=liferay-npm-scripts/v28.0.2`).

But we can make that easier for the user by accepting a shorter format
(eg. `v28.0.2` or even `28.0.2`), and then automagically doing the right
thing by reading the `version-tag-prefix` from the `.yarnrc`.

Test plan: I tried this in a bunch of scenarios to show that it behaves
reasonably.

For starters, in this repo, with:

```
../liferay-changelog-generator/bin/liferay-changelog-generator.js --version=$VERSION
```

Valid: `liferay-npm-scripts/v100.10.10+x-1`

    Wrote ./CHANGELOG.md 59835 bytes ✨

Invalid semver: `liferay-npm-scripts/v100`

    warning: Version "liferay-npm-scripts/v100" does not match expected pattern
    run again with --force to proceed anyway
    Error: Unexpected version format

Valid semver, missing "v": `liferay-npm-scripts/100.10.10`

    warning: Replacing unexpected "liferay-npm-scripts/" prefix with expected "liferay-npm-scripts/v" prefix in version "liferay-npm-scripts/100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 59823 bytes ✨

Misspelled package name: `liferay-npm-scriptz/100.10.10`

    warning: Replacing unexpcted "liferay-npm-scriptz/" prefix with expected "liferay-npm-scripts/v" prefix in version "liferay-npm-scriptz/100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 59823 bytes ✨

For comparison, with `--force` and that same bad name:

    warning: Version "liferay-npm-scriptz/100.10.10" is missing expected "liferay-npm-scripts/v" prefix
    Wrote ./CHANGELOG.md 59820 bytes ✨

Missing package name: `100.10.10`

    warning: Replacing unexpected "" prefix with expected "liferay-npm-scripts/v" prefix in version "100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 59823 bytes ✨

Missing package name, but with "v": `v100.10.10`

    warning: Replacing unexpected "v" prefix with expected "liferay-npm-scripts/v" prefix in version "v100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 59823 bytes ✨

Now, remove the `.yarnrc` and try again:

`v100.10.10`:

    warning: Removing unexpected "v" prefix from "v100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 59760 bytes ✨

`100.10.10`:

    Wrote ./CHANGELOG.md 59760 bytes ✨

`pkg/v100.10.10`:

    warning: Removing unexpected "pkg/v" prefix from "pkg/v100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 59760 bytes ✨

Without the `.yarnrc`, note how we strip off the "v", because none of
the tags in the repo start with "v", so we assume it is not wanted.

Now in a non-monorepo (eg. alloy-editor):

Testing with:

```
../liferay-npm-tools/packages/liferay-changelog-generator/bin/liferay-changelog-generator.js --version=100.10.10
```

    warning: Replacing unexpected "" prefix with expected "v" prefix in version "100.10.10"
    use --force to disable this coercion
    Wrote ./CHANGELOG.md 114051 bytes ✨

Note how in alloy-editor where all the tags start with "v", it correctly
infers that then "v" should apply.

Obvious disclaimer: if you feed totally random garbage into this, you'll
get unpredictable results, but that was always the case, and you should
always check the generated changelog before committing it.

Closes: https://github.com/liferay/liferay-npm-tools/issues/408