feat!: #148 Update to gulp 4
Because I have made the changes while exploring them it has been impossible to
commit them in split chunks, so I will explain what goes inside this commit:

1) I have created a `project` module to hold singleton values that before were
spread all over the code and were difficult to control and manage their state.

The `project` module contains references to the gulp instance, the options
passed by the user, the store (formerly `gulp-storage`), and the theme config
object (a section we place inside package.json to hold theme's metadata).

The `project` object also holds a reference to `runSequence`. This is a
method to be able to run gulp tasks in series which is no longer needed in gulp
4 but, because all the solution is based on it, I decided to keep until we
decide what to do with gulp 4 support.

2) I have changed implicit requires of gulp plugins to explicit ones to make
dependencies traceable.