chore: Add release:snapshot script
The release-snapshot code is copied from JS Toolkit. Maybe we should make a
common lib in the future but given that projects may have subtle differences and
that it's not so critical to have the release scripts aligned, we may leave it
like this for now (copy&paste) and decide what to do in the future.