fix: stop checkLinks.js from freaking out about codepen.io links
As explained in the related issue, Cloudflare rejects our link checks
with a 403 and an HTML page that asks us to fill out a CAPTCHA. I
checked to see if some lighter-weight check like an OPTIONS request
avoided the issue, but it doesn't.

So, grabbed some headers from the Firefox network pane and this seems to
be a reasonably minimal not-robot-like set that lets the requests
succeed.

Closes: https://github.com/liferay/liferay-frontend-guidelines/issues/133