chore: fix release process
Because `yarn ci` doesn't exist at a project level, the release failed.

We have decided to move the `yarn ci` to the top level so that it is only done
once per release (given that it lasts ~30s) and there's no point in repeating
the same process three times.

On the other hand, we have added a `yarn ci` script per project for comfort, not
related to the release.