chore: tag releases per package
We tried to use liferay-js-publish to release all projects at once but that's
not possible now.

Thus, we have decided to change to independent versioning in this repo. This
decission is made because the life cycles of each package are different so it
makes sense to use independent versioning.

Also, we don't want to revert what has been done for liferay-js-publish because
the situation is at least as good as in liferay-npm-tools and it is a step
towards full release automation.

The only things missing as of now are:

1) Automated changelog launch
2) Some safety checks to make sure that inter-project version references are
correct.

Those will be addressed in the liferay-js-publish in the future.