feat: teach "liferay/import-extensions" to check re-exports as well
Note that two of the test cases are disabled (for now, will fix in the
next commit) because "espree" doesn't handle "export * as name" syntax.

We are using this syntax in one place in liferay-portal right now:

https://github.com/liferay/liferay-portal/blob/1fd785b60f238666c35609f77c2a0d7c605eefdb/modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/index.js#L23

And it works there because we use "babel-eslint" as a parser. It would
also work in Clay where we use `@typescript/eslint-parser`, provided we
keep the dep up-to-date.

Closes: https://github.com/liferay/eslint-config-liferay/issues/164