feat: run tests against babel-eslint parser as well
This enables me to uncomment the tests that espree would choke on. Note
that not even the latest release of espree handles this syntax, so
switching to babel-eslint (like we use in portal) or
`@typescript/eslint-parser` is the only way to parse this syntax.

Nevertheless, I don't want to disable the espree tests entirely, so I
extended the MultiTester to run all three parsers, and added support for
a "skip" property that lets us skip the tests that don't run right now
under espree.